Added !exit command to stop the script via chat
A command you can use in chat to stop the script gracefully. The bot
will confirm via chat message & a simple timer code will print in your
cmd/interpreter window the desired seconds before ending the script.
NOTE: You must define the "sec" variable! Mine is set as "2" & imported
via cfg module, but can be set in the bot's module if desired (This is
for the timer code). Also, this does not log the account out directly
(It will just log out due to connection loss).